server/requirements.txt
async-timeout==4.0.3
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
python-dotenv==1.0.1
redis==5.0.6
requests==2.32.3
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.2
